update for tracks
Added track api endpoint. Updated readme to match. also added /with-tracks for albums endpoint and /with-albums for artists. Now a request to http://localhost:5000/api/albums/[id]/with-tracks or http://localhost:5000/api/artists/[id]/with-albums will bring up all associated tracks for an album or all albums associated with an artist respectively